Update IME parent onConfigurationChanged
When an app is moving between fold and unfold state, screenWidth will
change in DisplayContent#onConfigurationChanged() and depending on if
InputTarget is not-resizeable window, IME should no longer be attached
to the app to use full width.
We should update IME parent by calling updateImeParent() and it shall
take care of handling all such cases.

Fix: 183484673
Test: atest DisplayContentTests
Change-Id: I9e892c4c320c61d4a9ceefaf309e76b816c73388

Former-commit-id: 5fcf28e8bd5df056c12d1bc151d275ac398e07d2