Refactor QSTileView hierarchy
This CL collapses the hierarchy QSTileBaseView - QSTileView -
QSTileViewHorizontal into QSTileViewImpl. This can be done because now
there's a single type of view for the tiles. The benefit of this is that
we do not have to hack our way to undo things that are done in parent
classes. As part of this remove a lot of unnecessary files.

As part of this, bring some colors/sizes up to spec and make sure that
the sizes are reloaded in config changes.

Test: manual
Test: atest com.android.systemui.qs
Fixes: 187061459
Bug: 186057842
Change-Id: I63dbe8fa2b44833c11a486e84195e33e170c6f58

Former-commit-id: 9c8de7bdf10081516ab8a2a5bb130173797507f3