Animate app launches from lockscreen
This CL animates the app launches (from notification, media, etc) on the
lockscreen when the lockscreen can be dismissed (i.e. when there is no
security or when SmartLock/FaceAuth is enabled).

Note that in order to try this CL, you need to enable remote keyguard
animations:

$ adb shell setprop persist.wm.enable_remote_keyguard_animation 1
$ adb reboot

See b/184726377#comment20 for before/after videos.

Bug: b/184726377
Test: Click notification/media/QS from lock screen.
Change-Id: I27e616c1549676f6d1aca53b71efc18db8eec77a

Former-commit-id: f90ca12d50b3b4aed3ffc32af68e1e1bdb241043