Fix hard dragging the window magnification from the bottom of the screen
The window magnifier is hard to drag when it is at the bottom
of the screen. This is due to that launcher is detecting
swipe-up system gesture.

To fix it, we set the system ui state flag to true when
the window is overlapped with the gesture detection area.
We also update the gesture dection area via OnApplyWindowInsetsListener
Once the area is changed, we will update the flag again if necessary.

In the testing part, we use real WindowManager to verify the
functionality.

Bug: 179648683
Test: atest com.android.systemui.accessibility
      manually test and use the following command to check the flag
      adb shell dumpsys activity service TouchInteractionService
Change-Id: Ia09bd26b3eb164790614c173594d997d45f4ecdc

Former-commit-id: 0b0634c4a9c132a6061387a03b9e6690a17ac521