Throttle the repeated silent installation requests
To avoid a non-system installer repeatedly silent updates and
'denial-of-service' against another app on the device. This CL
tracks for all the silent updated installs where the
`requireUserAction' is not required. And fall back to user
action required if a repeated silent update is requested within
the throttle time.

Bug: 185878964
Test: atest SilentUpdateHostsideTests
Test: atest StagingManagerTest
Test: atest PackageInstallerSessionTest
Change-Id: I4ea287bf66d8661eec2676b604f9109f6f9add28
Merged-In: I4ea287bf66d8661eec2676b604f9109f6f9add28

Former-commit-id: ec065a8333fff9f39d697877bee2663e39cbdb1f