Fix measure logic so that priority children can have more than 1/N space.
The priority previously just put the view at the front of the list, which ensured that super long user-provided actions couldn't cause the priority action at the end to be truncated.  However, this still meant that a priority action at the start whose text would have it be longer than 1/3 of the action space would be capped at 1/3 instead of taking on some of the user-defined child's space.

Bug: 187533622
Test: Post Notify2 app's Call style w/ Screen action; notice that "Decline" is not truncated.

Change-Id: Ia18b2dc366ba7cac13ed845acac59d89206e7d5e

Former-commit-id: c6d567d2399a2e6514f961c04416c8e1f263a428