Rotation-invariant status bar layout calculator
This CL implements a handful of methods that enable us to pre-emptively
calculate the position of the status bar, taking into account
DisplayCutout and rounded corner content padding.

The status bar is very special in that it always sits at the top of the
device, and has a known height, only taking into account window insets
with regard to its content layout. Therefore it can be known in all
orientations what the positions will be before we get a chance to lay
out the views on a configuration change.

Still to do: use this to inform Window Manager APIs, and use this
information to rework the privacy dot view controller not to require a
late publication of the status bar location

Test: manual
Bug: 187973222
Change-Id: Ieeba175065c522a0616f109b3670b752faef7f2a

Former-commit-id: 98dd5315aa9022fb4bb9d72321c41c377742e47a