Don't clear size compat mode on fold/unfold.
With ag/14036667, the displays are turned off during fold and unfold transitions. While screen is off, ActivityRecord#mVisibleRequested is set to false and mCompatDisplatInsets is cleared in onConfigurationChanged > restartProcessIfVisible > clearSizeCompatMode. As a result, unresizable activities don't enter size compat mode after fold/unfold as they should. To address this checking that visibleIgnoringKeyguard is set to false before clearing mCompatDisplayInsets.

Fix: 187338605
Test: atest SizeCompatMode
Change-Id: I75ea5bba106d04c849e6b7ab446f77fcd8928981

Former-commit-id: 437319e18b119e3b0c1d351e200aa92bbe8eb04d