Cleanup the PM Computer class
Bug: 188048343

The Computer exposes a number of interfaces that are no longer needed
outside the computer itself.  The declarations are removed from the
interface.  PM trampolines are also removed.

If the method is never locked then it is made private in
ComputerEngine.  Otherwise it is made protected in ComputerEngine (so
it can be locked in ComputerEngineLocked) and it is made final in
ComputerEngineLocked.

Test: atest
 * CtsContentTestCases:IntentFilterTest
 * CtsDynamicMimeHostTestCases
 * CtsRoleTestCases
 * FrameworksServicesTests:UserSystemPackageInstallerTest
 * FrameworksServicesTests:PackageManagerSettingsTests
 * FrameworksServicesTests:PackageManagerServiceTest
 * FrameworksServicesTests:AppsFilterTest
 * FrameworksServicesTests:PackageInstallerSessionTest
 * FrameworksServicesTests:ScanTests
 * UserLifecycleTests#startUser
 * UserLifecycleTests#stopUser
 * UserLifecycleTests#switchUser
 * FrameworksServicesTests:WatcherTest
 * android.appsecurity.cts.EphemeralTest
 * android.appsecurity.cts.InstantAppUserTest
Change-Id: I42a70ee2d7a1e1c562db482c30260361afc5b617

Former-commit-id: 457d516dc00ba7fad59be72aee50c6f2554b2b26