PiP: Move resetting mAllowTouches into scope of MenuView
Previously we used to reset mAllowTouches backed on callbacks from
TaskOrg's animation finish, but there is a race where it's possible the
callback from the animation finish happens before the menu state
updates, which causes the callback to no-op. Instead, since we only
disable allow touches during #showMenu, just keep it in the scope and
re-enable it when the same animation has ended (or cancelled).

Bug: 188464003
Test: Move PIP to smaller, tap (to cause it to grow larger), and tap on
an icon, and expect icon to work and launch Settings, etc.

Change-Id: Iea3fff3c279dcdfa7eb922cb59302ea53ac618fd

Former-commit-id: ea4f7e5b2ab1cc36e71aae9c7d61e350bc833154