Translate: Ignore duplicate translations.
This can happen if UiTranslationManager#startTranslation is called more
than once for the same view. If the same TranslationResponse is
returned, it can simply be ignored.

This is preferable to ignoring duplicate ViewTranslationRequests, since
the Response is already cached. The TranslationService can cache
translations, so duplicate requests isn't an issue.

Ideally, the platform intelligence shouldn't issue multiple requests for
the same view.

Bug: 144292180
Test: atest CtsTranslationTestCases
Change-Id: I0bba4a19284a61a9c7a5b79c71b4f42ddffe4e32

Former-commit-id: 02fe90bb4202123bc21489e4f19fb3427b2f5745