Add an extra colored layer when animating launches (1/2)
This CL adds an extra layer with the same color as the app background
color when animating an app launch (e.g. from notification, QS, media,
etc). This allows to avoid a "flash effect" that could be caused when
the background color of the expanding view is bright, which is often the
case for QS tiles.

See b/188185381#comment5 for before/after videos.

Bug: 186189658
Test: Long press a quick settings tile
Change-Id: Ifc0cda3b03795549ddd7c333a9302e7ea298ac6a

Former-commit-id: ed10f0ab6f8124b12681d36ab3b260ff342b2cd6