Fixed an issue where the keyguard media visibility was wrong
Because we were overriding the visibility to avoid some bugs,
this accidentally had the effect that hidden media players would
become visible. These visibility overrides were masking an
issue where the hostview wasn't attached to neither of the
containers, leading to the host not receiving any updates.
The mediaHost was waiting until being attached to start
listening, which is now also changed to avoid similar
future issues.

Fixes: 188530838
Test: atest SystemUITests
Change-Id: I8f5aad143713c057cec068f42b5f68cffc370861

Former-commit-id: e9587bb557aaf3513dcf8d5ba4cbaeb46460025a