Test for null package in isEnabledAndMatches()
Bug: 188174037

Avoid an NPE is isEnabledAndMatches() by returning false if no package
is found matching the component name.

This does not address the more general problem uncovered by the bug,
which is that a thread that has multiple calls into PM on the stack
might end up executing with different snapshots.

Test: atest
 * CtsContentTestCases
Change-Id: Ib4e929a62edf3393e0177f5d85c9b3fe1e0697d8

Former-commit-id: ff9a01834e228249d39307c943b1f2e8b4b379a4