Assigns the default value of the cache magnification mode
The magnification switch button is not visible because the
image source is not set when calling showButton().
Because the default value of cache magnification mode is
the same as the current magnification mode.
And the check in showButton() blocks setting the image source
To fix this for fullscreen mode, the default value is replaced
ACCESSIBILITY_MAGNIFICATION_MODE_FULLSCREEN with
ACCESSIBILITY_MAGNIFICATION_MODE_NONE.

Bug: 188857483
Test: atest MagnificationModeSwitchTest
Change-Id: I36eb7b20f94adbab49f006627e60badeb0b0f02f

Former-commit-id: 690c6914c4529f8feb48db7e6593270b46472f46