Fix potential exceptions in CallDiagnosticService and missing locks.
There is a possibility of a call receiving a details update when it has
already been removed from the CallDiagnosticService.  Handling this case
by just ignoring the details update (since it isn't relevant as the call
is now removed).
Also noted there were a number of places where access to the shared
maps was not synchronized as it should be.

Test: Run manual tests and note no more CDS crashes from SCONE app.
Fixes: 188939131
Change-Id: Ic1f96fd60a70432bef0daaa2e9729c66401ba4ac

Former-commit-id: 07d3736f5daf61b8a10b8a22f4bfb7c8ef87b635