Fixes issues with multiple ringer rows (which also addresses a few RTL/landscape issues).
The main change here is that the volume rows draw their own rounded backgrounds, aside from the last row, which draws no background since it's already on top of the main dialog background. The ringer icon and the ellipsis also now handle drawing their own rounded backgrounds, rather than having the container handle it. The container needs to draw the background for the main ringer row for Z-ordering reasons.

There's also some drawable trickery to get the landscape drawer expansion to work, so that it has a rounded background behind it when expanded.

Test: in portrait/landscape, with multiple rows (by using talkback/phone call), and in RTL
Fixes: 187202281
Fixes: 188038844
Fixes: 182833291
Change-Id: If969d29780f99b1454edd24e85e51a5d0ebe5d6d

Former-commit-id: 69a17f570526bc4df8efad3aa0c9e3afe83fbb59