Use PWLE for waveforms and add ramp down to zero
Add new configuration for ramp down duration (defaults to zero) to be
used by devices that require a soft ramp down to zero amplitude
between two segments when the change is abrupt to zero.

The waveform timings will be respected, so the ramp down duration might
be reduced if the waveform segment with zero amplitude is shorter than
the configured ramp duration.

This does not add a ramp at the end of vibrations. This will be done in
a follow up.

Bug: 188431691
Test: StepToRampAdapterTest
Change-Id: I9af8c653462defaa1fc26a4fe60077a63831f3fe

Former-commit-id: 11e3bb7af7813548a0e2cfa3a80f44a00ce0b4cf