Implement a callback to report HotwordDetectionService status
When the application uses VoiceInteractionService to create a
SoftwareHotwordDetector, we need to report the HotwordDetectionService
status to the application, but we pass null callback into the
updateStateLocked function, it means that we will not receive the
HotwordDetectionService status.

In this change, we implement IHotwordRecognitionStatusCallback and
pass the callback into updateStateLocked function. When we receive
the HotwordDetectionService status, we will pass the status to the
application.

Bug: 183239093
Test: atest CtsVoiceInteractionTestCases
Test: atest CtsVoiceInteractionTestCases --instant
Change-Id: I1a1b8abe5c7e82eddaa2f0a83f54110df74d1efb

Former-commit-id: 645623e732c27e21bd1efdee04dc118e18ad6464