Force relayout if sync operation is requested explicitly
The initial value of WindowState#mRedrawForSyncReported is false.
Then even if mSyncState is none and mPendingDrawHandlers is empty
and Task#mMainindowSizeChangeTransaction is null, it still forces
client to relayout.

This change aligns the condition in updateResizingWindowIfNeeded
that also checks shouldSendRedrawForSync. So if there is no
explicit sync request, there won't be an extra relayout.

Bug: 189915038
Test: CtsWindowManagerDeviceTestCases
Test: Check trace of adding a new simple window. There should be
      only one traversal.
Change-Id: Iad10034e29b9db50ec4221ac1a25f26ce6464393

Former-commit-id: 6a0777a1d778f6b469722b25a0e7645a2adce525