Fading out keyguard only content instead of occluding them with the scrim
The scrim is also used on the lock screen and should therefore be
behind the keyguard elements. Instead, we now fade out the views
on top of the scrim to achieve the same visual effect.

This also fixed an issue where the inset on the scrim was wrong in landscape

Fixes: 188660106
Test: drag down on lockscreen, observe content fading out
Change-Id: I72e6b964dde5ad3c579a0db71d7914f3e47eab6c

Former-commit-id: aca32441ea53a03319661f9cf5edb74c1f6f0772