Create platformtestprotos.
This target includes a subset of protos in platformprotos, and will
replace it in TradeFed. It is created as a separate file so that it can
have its own set of owners. The owners will be updated in AOSP, but this
change has to be submitted first due to this target requiring one
additional dependency compared to AOSP.

Bug: 181921006
Test: The target builds.
Change-Id: I82386a6c8a5e6ddf42d8377148001f4ede2f297f

Former-commit-id: d61699e86c77fe0d5998f8ced019affb35800645