Fix LegacyGlobalActions broadcast receiver
By default CLOSE_SYSTEM_DIALOGS broadcast is sent only for current user.
Current user is user 10 on devices with headless system user enabled.
LegacyGlobalActions broadcast receiver is registered with system user (user 0),
so the broadcast is not received.
It is proposed to use registerReceiverAsUser with UserHandle.ALL when
registering receiver to be able to receive broadcast.

Fixes: 182842046
Test: atest CtsAccessibilityServiceTestCases
Change-Id: I10405769ba6426c7bbcd31e09e1cd4c99585665b
Merged-In: I10405769ba6426c7bbcd31e09e1cd4c99585665b

Former-commit-id: 27be1ccfeb899ff4899968a4de3244724a193205