Fix margins around tiles in QSCustomizer
When tiles are moved, only those that are marked as modified will be
forced to update the item offsets, with the others maybe never getting
the chance. Given that moving may make tiles change columns and
therefore their margin, this is not good.

Instead, have the LayoutManager always poll the MarginItemDecoration for
the correct margins for elements that are not TextView (tiles), then set
that as the margin.

Also, fix margins for RTL.

Test: manual
Bug: 188640649
Change-Id: I4aaef4887a6f50e5070c764048b67792dec3a35c

Former-commit-id: 09e9a3be0b91a4ce70144b04d4ef66bf5bee575a