Make ApiLint's new_since arg the merged txt file
ApiLint tries to only lint "new" APIs, i.e. those introduced since the
last frozen API. Using android-non-updatable's last frozen txt does not
work well for this, as it is incomplete API surface on its own, so
metalava cannot see all APIs that were present in the previous API
(such as methods inherited from ART), and so makes the wrong decisions
for which APIs are new.

Bug: 171506470
Test: m checkapi
Change-Id: I00d25f2ae0ba8f0ee6fa8564d86642efd74e0d44

Former-commit-id: 4f90aa5148544f3bf9edebb68e02966f69a3ae03