A space-efficient 2D matrix - clean up
Bug: Bug: 188447813

Convert WatchedSparseBooleanMatrix to pack bits into ints instead of
bytes.  This eliminates a lot of casts and it is probably slightly
faster.

Force the control information arrays in the matrix to resize in step
with the values array.  The default behavior of GrowingArrayUtils is
to double the size of an array when it needs to grow, which is
unnecessarily large.  Also, the control arrays are shrunk along with
the value array.

Use the ArrayUtils methods for array allocation.  This is consistent
with classes in android.util.

Test: atest
 * CtsContentTestCases:IntentFilterTest
 * CtsDynamicMimeHostTestCases
 * CtsRoleTestCases
 * FrameworksServicesTests:UserSystemPackageInstallerTest
 * FrameworksServicesTests:PackageManagerSettingsTests
 * FrameworksServicesTests:PackageManagerServiceTest
 * FrameworksServicesTests:AppsFilterTest
 * FrameworksServicesTests:PackageInstallerSessionTest
 * FrameworksServicesTests:ScanTests
 * UserLifecycleTests#startUser
 * UserLifecycleTests#stopUser
 * UserLifecycleTests#switchUser
 * FrameworksServicesTests:WatcherTest
 * android.appsecurity.cts.EphemeralTest
 * android.appsecurity.cts.InstantAppUserTest

Change-Id: I8427905361e1471c91433c77628aa34b12b1044d

Former-commit-id: 44b3697cd2a1f2548018e2feefa6bbb5d9cc6f75