Execute app transition if starting window is drawn
For non-cold launch, the transition is executed after scheduling
to start/resume next activity. But for cold launch, the activity
will be started until its process is attached. Because transition
should be able to run if the starting window is drawn, it can
be executed earlier to have better response time. That may reduce
transition delay up to 40ms.

Bug: 161781889
Bug: 182836977
Test: Cold launch app with starting window. Check the trace
      that AppTransitionReady happens before attachApplication.
Test: atest CloseImeAutoOpenWindowToAppTest

Change-Id: I66b9a6ef4e14289ddbecb4971620981a9eca4afa

Former-commit-id: f80e196f4d4ceb9b0cdf85d16453e3639186f1e4