Create splash screen if activity launched with FLAG_ACTIVITY_CLEAR_TASK
When an activity launched with NEW_TASK and CLEAR_TASK, there will
finish the original activities and reuse the same task to create a new
activity, so there will need to wait for the new window drawn to
trigger wm transition, which slow down the launch response. And because
this condition should be a kind of warm launch, there should create a
splash screen for it.

Bug: 187808226
Test: atest ActivityRecordTests SplashscreenTests
Test: manual verify by winscope
Change-Id: I263edcd75fcd21999b9ceb9fac8bfcf8419539a3

Former-commit-id: ed9c44a27331ff535611d715ccea5b3edcfffc6c