Add log for IPC failure
Sometimes we will fail to schedule TopResumedActivityChangedItem because
of IPC failure and will cause other problems.

Client side will ignore the inconsistent state in non-debug build and
there is not much can do on the server side. Just add a warning log so
that we can easier debugging and confirm the root cause if issue
happens again.

Bug: 188843741
Test: atest ActivityRecordTests
Change-Id: I6e45f7bbd69384a3b734f80df4bc7c96ffb85b0d

Former-commit-id: 17f3aecf8767035d79d872e3b30315df1f21c8e8