Fix XML serialization of WallpaperColors
WallpaperColors is serialized to disk for use the next time the device
boots. WallpaperColors wasn't serializing the quantizer results, just
the 3 "main" colors, so it was unable to identify more than a few
colors, using the AOSP algorithm.

Bug: 186203148
Test: Reboot device without patch, verify only a couple colors are
available, or none if the AOSP algorithm for selecting main colors
differed sufficiently. Verify that with patch, rebooting the device
shows the same color options in the same order as it did at runtime.
Change-Id: I881a2cfe5b7dda5b7734ce3fda7275dc5544ce0c

Former-commit-id: 3a00015faab1b5a4eab566abd3622cd43e3612b0