Allow transfer starting window to windowDisableStarting activity.
Fix a black screen case when a trampoline activity finish itself
after it start another activity in the same task.
It was not allow to create a starting window on second activity
if it has declared as disable starting, but if the starting
window already exists, we should allow the transfer from the
previous activity.

Bug: 190671916
Test: cold launch PreDeferredSetupWizardActivity from notification.
Test: atest ActivityRecordTests
Change-Id: I75d89dc4152818b145951cbb9489d8bc2aa54340

Former-commit-id: 9ecf891254286397982cac1dc9bf4c5cd9c8e007