Separate trace of trampoline launch with different packages
The trace of app launch is "launching: $packageName". Previously,
if a caller uses its trampoline activity to launch another app,
the trace name will still show the package of caller. So now if
the packages are different, start a new trace for it.

Also update last transition info so it won't miss the fully-drawn
from the last launched activity.

Bug: 191765029
Test: Use Searchbox to launch other apps. The trace should
      show two slices including launcher as trampoline and
      the target app.

Change-Id: I3f2f19323089385d262a258f70b510f18e9943c3

Former-commit-id: 91cc2b25977947594f2f4bcdfd4c2ec6bcccb9c7