Fix screenshot to corner animation
We now offset the location of the screenshot preview based on its
location in the screen, but we were setting the scale in between
making the "location in window" and "location in screen"
measurements, which meant that the preview got offset to the wrong
place.

This change moves any changes to the actual views into the
onAnimationStart method, to logically separate the animation setup
calculations from the animation itself.

Bug: 191894245
Fix: 191894245

Test: manual

Change-Id: Idc3d7654268d277f7416a025952ee6654e1a0630

Former-commit-id: f2c1c3cba8acf04bbf5fe89758f926767589045c