Fixed an issue where notification dismissals wouldn't animate
For notification groups with a single child, where the parent
is suppressed, no notifications were generated for the dismissal
because the stackscroller thought that a child transfer was in
progress.
This also fixes a case where we accidentally triggered a resize
animation because remoteInputs were removed even though they
haven't changed their visibility. This was masking the above
behavior from manifesting always and only showed for notifications
that had no remote input.

Fixes: 187291379
Test: add group with a single child and no remote input, dismiss, observe animation happening
Change-Id: Iab4f6697feb393c4e1cc8b5f546c9daa110f3230

Former-commit-id: 59d8322d5e175f354e2b0c6187e7d1a6c1f840ed