Enable sensor use notification on TVs.
Notification is shown as an alternative to the bottom sheet / dialog
when the a blocked sensor is being accessed.
Only show the notification briefly to notify the user but don't keep it.

Test: manual - change code to use notification instead of bottom sheet,
block sensor (camera or mic), start appop for that sensor, notification
shows up but isn't in the notification list afterwards.
Bug: 186720349

Change-Id: I08d860812d14be4b36b84cf8869fe034941f5268

Former-commit-id: 6c4798b19d2b1faf977032ae320ce49ac376e458