Remove activity from unknown app if starting window is relayouted.
App transition would need to wait for unknown activity become resumed
or relayout if it was start when lock screen on, and there may take a
long time to wait for an app relayout. But actually there will create
starting window as soon as keyguard is unlock, so there can listen for
starting window drawn as a sign to remove the activity from unknow app
list. There can save hundards ms if there are launch a large app.

Test: launch app from lock screen with/without set credential.
Bug: 187808226
Change-Id: I088a0f88c310ca8f2d886f078a318bbaca21f6c3

Former-commit-id: 6d37a5ea605d89f7ad1743997c43d964c0b2c1c1