Turn off screenshot transition into long screenshots for rotations
If the user rotates the phone between the initial screenshot capture
and the long screenshot capture, the views will be completely
different and it doesn't make sense to try to transition from the
preview into the long screenshots activity.

This change just turns off the preview transition for this case
(though we still show the scrim and fade it out during the
transition).

Bug: 192344739
Fix: 192344739
Test: manual
Change-Id: I5951c074ad8ddb3957a76df963979cf47c987a66

Former-commit-id: 9c759448ee1ac9062c0e953db593a08c742a9bf0