Add support to slowly bring the vibrator down in VibrationThread
Add support to slowly bring down the vibrator amplitude when a vibration
ends at a positive amplitude or when it is cancelled while playing at a
positive amplitude.

This uses the same duration configured to ramp down PWLE and step
waveforms, and the duration is not considered part of the user defined
vibration.

Bug: 188431691
Test: VibrationThreadTest
Change-Id: I642e898c3b13b2c841543e870275e40df4d00b81

Former-commit-id: 3c729321dc64bbeb5b60d51c25491baf06ea8a80