Fix clipping of SecondaryLabel on animation
With this CL, the label container in collapsed tiles will be measured
with vertical MeasureSpec.UNSPECIFIED. That will make sure that the
labels take as much space as necessary.

In collapsed, the secondary label is not seen, but it's measured after
the available space is given to the first one and results in a line
that's not high enough. Using UNSPECIFIED makes sure that it gets as
much space as it needs. The total height of the label container is not a
problem since we know that the visible contents in the tiles will always
fit (after accounting for paddings).

Test: manual, also using high text size
Fixes: 187485111
Change-Id: Ida0d05a44d0722e85dc77018b5b2eabe334d47b2

Former-commit-id: bb566ca42ea1c207c8cc5174a4f445fbbb8d1029