DO NOT MERGE Fixed the padding of quick settings
Quick settings was accounting for double the padding / or not enough
in most cases. The padding is now only on the quick settings itself
and the various workarounds in the build are reverted.
Media and the horizontal layout would obtain negative margins for
example, which where just cancelling themselves out.

Fixes: 192937466
Test: add media / no media, observe proper 24dp padding below
Change-Id: I635afaaf9185c32a9f04851453ac7c0d20612c11
Merged-In: I635afaaf9185c32a9f04851453ac7c0d20612c11

Former-commit-id: a501b9b9cb394f42b4ceaede8586f438a8f94e6e