Camera: Restore FastNative annotation.
FastNative annotations were removed from native methods since it can't
be used with 'synchronized'. These methods need not be synchronized.
As per comments in ag/13184856, when the annotations were added first
effect was that ART did the fast native calls but because synchronized
fast native methods are not supported it was not doing the synchronized
nor the unlock at exit.So FastNative but not synchronized.

Removing synchronized and restoring FastNative annotation to optimize.
Did sanity testing like open/close camera,taking photos,clear all apps.

bug:175396433

Test: Camera ITS scene0 Pass, All Camera CTS tests Pass.
Change-Id: Ib7acd363120ee0bf2a1dea92628b4697f3b211bb

Former-commit-id: 99ebd95ef140a14185b729e419abe19c6dc0a678