Fix VibrationThread with repeating waveforms of positive amplitudes
The VibrationThread is using a fixed duration of 1s to turn on the
vibrator for a repeating waveform that has all amplitudes positive.

This does not work if steps are longer than that, as the vibration will
stop prematurely and the thread will wait until the next step to turn
the vibrator back on, causing a noticible break.

Fix: 193206875
Test: VibrationThreadTest
Change-Id: I7ac3e397d588ac44dde5430f8fea852a283b4ae8

Former-commit-id: 0484b7e98ba11b59d0c150638a2382ffeecce943