Fix issue where screen off is disabled after expanding then collapsing the shade.
The root cause is that we only get a callback to update whether we want to control screen off when the shade expansion changes, not when the isExpanding value changes. Sometimes, this value can still be true briefly when the callback is triggered.

We can safely fix this by checking isFullyCollapsed rather than isExpanded || isExpanding. isFullyCollapsed is updated at the same time as isExpanded, so it will be accurate when the callback is triggered.

Fixes: 194212105
Test: expand the shade, collapse it, press power button, see animation
Test: expand shade, press power, don't see animation
Test: press power while in the middle of expanding or collapsing the shade, see no animation
Change-Id: I0396aa4ee6408820d700ac23bea04897f37c4443

Former-commit-id: cd3c4ec4260dd8c70c046ce0e6a2c7a5f6c1cc54