Update mic/cam quicksetting toggles when user changes
On a secondary/guest user the mic/cam toggles show for user 0. We want
the user-free calls to reflect the state of the current user so things
like the UI is accurate. Before this change the user free calls always
checked for user 0 since the context was cached in SensorPrivacyManager.

This change also fixes some issues with removing listeners.

Test: use toggles on main, work profile, and secondary user
      atest CtsSensorPrivacyTestCases --user-type secondary_user
Fixes: 178751362
Change-Id: Ieb90ea7ba4074446eb01d1093e280ac7018f9e73

Former-commit-id: 8f7f7ab98aec9cdd61718e352dd821dc3079bd2a