Fix test harness mode
There's legacy request that when OEM unlock is enabled, PDB service
format the PDB in the onStart to make sure the fastboot -w
can wipe the PDB for devices that do not have recovery OS. In the
past, only the header is cleaned up so test harness can still peek
the data in the PDB. b/185369514 is an issue due to the incomplete
format so an explicit clean up is required. This CL makes it half-format
to support both use cases.

Bug: 192268220
Bug: 18644051
Bug: 185369514

Test: Test harness
Change-Id: I290c5b2fde9f1337b41337795e4c2fbcb9e95f6e

Former-commit-id: 574f47af6514f346c5cb9712cc160e2c4b98afd5