DO NOT MERGE [QPR1] Fix QS having a white when opened over bubbles
mBehindTint now controls the background of quick settings.
I think this wasn't the case in R when the  bubble state
was written & it was never updated appropriately whenever
that changed.

Seems like most of the code in the unlocked state should
apply to bubbles so just copied it in and updated the
test.

Test: atest ScrimControllerTest
Bug: 191338071
Change-Id: I90fa26760d346a67d42b22d8e562d5b4c66f3017

Former-commit-id: 55479777dd42067ff5e8830454c18f8bb817757d