Improve restore of QS_AUTO_ADDED_TILES
In order to prevent race conditions and effects happening before restore
happens, this CL implements the following changes:

* QS_AUTO_ADDED_TILES is restored right after QS_TILES
* Writes of QS_AUTO_ADDED_TILES are marked as overrideable by restore.
That way, the restore will happen even if tiles are auto-added before
the restore happens.
* If/when the restore happens. Any tiles that had been auto-added and
manually removed in the source device are removed from the set of
current tiles.
* If/when the restore happens, QS_AUTO_ADDED_TILES is resolved to be the
merge of the value before and after restore.

Test: atest SystemUITests
Test: manual restore
Fixes: 168687572

Change-Id: Iad151e8310f1a5b099125d20bb673703030501b6

Former-commit-id: e2cce552ede6c86e65a205f357863ab3151b5302